+++ pkg/epp/flowcontrol/framework/plugins/ordering/fcfs_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package intraflow
+package ordering
 
 import (
 	"testing"
